requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
uvicorn[standard]==0.15.0  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.1.3   # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
# django-allauth==0.45.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.0.0  # https://github.com/jazzband/django-redis

# Local
# -------------------------------------
phonenumberslite==8.12.36
django-admin-interface==0.22.2
ShopifyAPI==9.0.0
python-dateutil==2.8.2
django-phonenumber-field[phonenumberslite]==7.0.0
